changed texrgbscale to int rather than float, this allowed me to change all remaining glTexEnvf calls to glTexEnvi
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@1799 d7cf8633-e32d-0410-b094-e92efae38249

+++ gl_backend.h
@@ -21,7 +21,7 @@ typedef struct
 	int tex[MAX_TEXTUREUNITS];
 	float *texcoords[MAX_TEXTUREUNITS];
 	int texcoordstep[MAX_TEXTUREUNITS];
-	float texrgbscale[MAX_TEXTUREUNITS]; // used only if COMBINE is present
+	int texrgbscale[MAX_TEXTUREUNITS]; // used only if COMBINE is present
 }
 rmeshinfo_t;
 
@@ -36,7 +36,7 @@ typedef struct
 	int numtriangles;
 	int numverts;
 	int tex[MAX_TEXTUREUNITS];
-	float texrgbscale[MAX_TEXTUREUNITS]; // used only if COMBINE is present
+	int texrgbscale[MAX_TEXTUREUNITS]; // used only if COMBINE is present
 
 	// output
 	int *index;
